+++ src/read.rs
@@ -525,3 +525,38 @@ pub fn consume_header_while<R: Read>(
         Ok(None)
     }
 }
+
+pub fn consume_lines<R: Read>(
+    reader: R,
+    limit: usize,
+) -> io::Result<Option<(u64, RecombobulatedReader<R>)>> {
+    let mut line_reader = simd_csv::LineReader::new(reader);
+    let mut pos = line_reader.position();
+    let mut seen: usize = 0;
+    let mut header_opt: Option<Vec<u8>> = None;
+
+    while let Some(line) = line_reader.read_line()? {
+        seen += 1;
+
+        if seen <= limit {
+            pos = line_reader.position();
+            continue;
+        }
+
+        header_opt = Some(line.to_vec());
+
+        break;
+    }
+
+    if let Some(mut fixed_data) = header_opt {
+        let bufreader = line_reader.into_bufreader();
+        fixed_data.push(b'\n');
+        fixed_data.extend(bufreader.buffer());
+
+        let fixed_reader = Cursor::new(fixed_data).chain(bufreader.into_inner());
+
+        Ok(Some((pos, fixed_reader)))
+    } else {
+        Ok(None)
+    }
+}