+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * `xan behead` & `xan rename` do not normalize the output anymore to be as fast as possible.
 * The new SIMD CSV parser might not deal with CSV irregular cases the same way `rust-csv` did. In any case, `xan input` will still continue to use `rust-csv`.
 * `xan slice -B/--byte-offset` & `xan slice -A/--accumulate` are now mutually exclusive.
+* `xan input` has been overhauled.
 
 *Features*